+++ Neural_Network/Neural_Network.c
@@ -4,7 +4,7 @@
 
 #define MAX_SIZE 8000
 
-NeuralNetwork* create_network(int input, int hidden, int output, double lr) {
+NeuralNetwork* create_network(int input, int hidden, int output, float lr) {
 
 	NeuralNetwork* network = malloc(sizeof(NeuralNetwork));
 
@@ -202,7 +202,7 @@ void train_batch_imgs(NeuralNetwork* net, uint8_t* images, uint8_t* labels, int 
  
 		for (int k=0,j = i * IMAGE_SIZE; j < (i+1) * IMAGE_SIZE; j++,k++)
 		{   
-			IMG->data[k] = (double)images[j]/255;
+			IMG->data[k] = (float)images[j]/255;
 		}
 
 	    
@@ -228,7 +228,7 @@ void train_batch_imgs_epochs(NeuralNetwork* net, uint8_t* images, uint8_t* label
 {
 	int epoch = 100;
 	int epochs = size / epoch;
-	double NET_RATE;
+	float NET_RATE;
 
 	FILE* imageFile = fopen("./mnist_reader/mnist/t10k-images-idx3-ubyte", "r");
 	FILE* labelFile = fopen("./mnist_reader/mnist/t10k-labels-idx1-ubyte", "r");
@@ -255,7 +255,7 @@ void train_batch_imgs_epochs(NeuralNetwork* net, uint8_t* images, uint8_t* label
  
 		for (int k=0,j = i * IMAGE_SIZE; j < (i+1) * IMAGE_SIZE; j++,k++)
 		{   
-			IMG->data[k] = (double)images[j]/255;
+			IMG->data[k] = (float)images[j]/255;
 
 		}	    
 	    int index = *(labels + i);  //recupere l'indice de label active
@@ -323,7 +323,7 @@ Matrice* predict_network(NeuralNetwork* net, Matrice* IMG,int i) {
 }
 
 
-double predict_rate_network(NeuralNetwork* net, uint8_t* images, uint8_t* labels, int size,int choix) {
+float predict_rate_network(NeuralNetwork* net, uint8_t* images, uint8_t* labels, int size,int choix) {
 	int img_correct = 0;
 	// printf("\n***********************************\n");
   
@@ -333,7 +333,7 @@ double predict_rate_network(NeuralNetwork* net, uint8_t* images, uint8_t* labels
 
 		for (int k=0,j = i * IMAGE_SIZE; j < (i+1) * IMAGE_SIZE; j++,k++)
 		{   
-			IMG->data[k] = (double)images[j]/255;
+			IMG->data[k] = (float)images[j]/255;
 
 		}			// recuperer la matrice qui contient les activations de l'image de test
